+++ test/python/test_git_url_rewriter.py
@@ -27,3 +27,25 @@ def test_git_url_rewriter_git():
 
     r = rewriter.rewrite_url('github.com/steinwurf/gtest.git')
     assert r == 'git://github.com/steinwurf/gtest.git'
+
+
+def test_git_url_rewriter_ssh_git_at():
+
+    parser = GitUrlParser()
+    rewriter = GitUrlRewriter(
+        parser=parser,
+        rewrite_protocol='ssh://git@')
+
+    r = rewriter.rewrite_url('github.com/steinwurf/rely-python.git')
+    assert r == 'git@github.com:steinwurf/rely-python.git'
+
+
+def test_git_url_rewriter_custom():
+
+    parser = GitUrlParser()
+    rewriter = GitUrlRewriter(
+        parser=parser,
+        rewrite_protocol='git+https://TOKEN:x-oauth-basic@{host}/{path}.git')
+
+    r = rewriter.rewrite_url('github.com/steinwurf/gtest.git')
+    assert r == 'git+https://TOKEN:x-oauth-basic@github.com/steinwurf/gtest.git'