+++ NEWS.rst
@@ -7,6 +7,9 @@ of every change, see the Git log.
 Latest
 ------
 * Major: Make resolve symlinks relative.
+* Minor: Added support for git_protocol `ssh://git@`
+  (used when "cloning" with pip).
+* Minor: Added support for custom git_protocols.
 * Patch: Recurse dependencies of dependencies before the dependency.
 * Major: Change the default interpreter to python3
 * Patch: Fix resolve such that it won't run twice when invoked explicitly.